[ci skip] [WIP] Rebased and fixing builds

+++ src/jet/point_hash_grid_searcher2.cpp
@@ -249,7 +249,7 @@ void PointHashGridSearcher2::deserialize(const std::vector<uint8_t>& buffer) {
 
     // Copy simple data
     auto res = fbsToJet(*fbsSearcher->resolution());
-    _resolution = {res.x, res.y};
+    _resolution = {(ssize_t)res.x, (ssize_t)res.y};
     _gridSpacing = fbsSearcher->gridSpacing();
 
     // Copy points
